Project Creation
After activating virtual environment, installed django 4.2.7 (latest version) and started the project named Bookstore where we will make a database, store the books and related information. Also, we will be able to fetch those details on the webpage and by clicking on any book we will be redirected to another page which will display all the related data of that particular book.