Defining functions to fetch data
Created 2 functions 1 that redirects us to the html page named index.html which shows the list of books available.
The second function is triggered when we click on any book name on the index page and redirects us to another html page named 'details.html' which will display the details of that particular book that we clicked.